# openapi.yaml
openapi: 3.0.0
info:
  title: HerokuAPI
  version: 1.0.0
  description: External Service for Heroku Travel App
servers:
  - url: https://travel-app-experience-b7e9d856670c.herokuapp.com
paths:
  /api/paises:
    get:
      operationId: getPaises
      summary: Lista de países disponíveis
      responses:
        '200':
          description: Lista de países
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/regioes:
    get:
      operationId: getRegioes
      summary: Lista de regiões por país
      parameters:
        - name: pais
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de regiões
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/categorias:
    get:
      operationId: getCategorias
      summary: Lista de categorias por região
      parameters:
        - name: regiao
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/restaurantes:
    get:
      operationId: getRestaurantes
      summary: Lista de restaurantes com filtros
      parameters:
        - name: pais
          in: query
          schema:
            type: string
        - name: regiao
          in: query
          schema:
            type: string
        - name: categoria
          in: query
          schema:
            type: string
        - name: notaMinima
          in: query
          schema:
            type: number
        - name: pagina
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de restaurantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestauranteResumo'

  /api/restaurantesPatchedHeroku/{locationId}:
    get:
      operationId: getRestaurantesByLocationId
      summary: Detalhes do restaurante por ID
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes do restaurante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestauranteDetalhado'
        '404':
          description: Restaurante não encontrado

components:
  schemas:
    RestauranteResumo:
      type: object
      properties:
        nome:
          type: string
        nota:
          type: number
        categoria:
          type: string
        location_id:
          type: string

    RestauranteDetalhado:
      type: object
      properties:
        nome:
          type: string
        nota:
          type: number
        categoria:
          type: string
        location_id:
          type: string
        detalhesJson:
          type: object
          description: JSON completo de detalhes do restaurante